comfyui_nodes/requirements.txt
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
torch>=2.0.0
Pillow>=9.0.0
numpy>=1.24.0
